+++ internal/DateTime/DateTime.js
@@ -1,6 +1,5 @@
 import kind from '@enact/core/kind';
 import PropTypes from 'prop-types';
-import React from 'react';
 
 import componentCss from './DateTime.module.less';