+++ tests/screenshot/apps/components/Slider.js
@@ -1,5 +1,4 @@
 import Slider from '../../../../Slider';
-import React from 'react';
 
 const SliderTests = [
 	<Slider />,