+++ tests/ui/apps/LabeledIconButton/LabeledIconButton-View.js
@@ -3,7 +3,6 @@ import Scroller from '../../../../Scroller';
 import ThemeDecorator from '../../../../ThemeDecorator';
 import spotlight from '@enact/spotlight';
 import {scaleToRem} from '@enact/ui/resolution';
-import React from 'react';
 
 // NOTE: Forcing pointer mode off so we can be sure that regardless of webOS pointer mode the app
 // runs the same way